# pkg/guacrest/generated/models.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.0.1-0.20240123090344-d326c01d279a DO NOT EDIT.
package generated

// Defines values for AnalysisDependenciesParamsSort.
const (
	Frequency AnalysisDependenciesParamsSort = "frequency"
	Scorecard AnalysisDependenciesParamsSort = "scorecard"
)

// Error defines model for Error.
type Error struct {
	Message string `json:"Message"`
}

// PageInfo defines model for PageInfo.
type PageInfo struct {
	EndCursor       string `json:"EndCursor"`
	HasNextPage     bool   `json:"HasNextPage"`
	HasPreviousPage bool   `json:"HasPreviousPage"`
	StartCursor     string `json:"StartCursor"`
}

// PaginationInfo defines model for PaginationInfo.
type PaginationInfo struct {
	PageInfo   PageInfo `json:"PageInfo"`
	TotalCount int      `json:"TotalCount"`
}

// Purl defines model for Purl.
type Purl = string

// PaginationSpec defines model for PaginationSpec.
type PaginationSpec struct {
	After  *string `json:"After,omitempty"`
	Before *string `json:"Before,omitempty"`
	First  *int    `json:"First,omitempty"`
}

// BadGateway defines model for BadGateway.
type BadGateway = Error

// BadRequest defines model for BadRequest.
type BadRequest = Error

// InternalServerError defines model for InternalServerError.
type InternalServerError = Error

// PurlList defines model for PurlList.
type PurlList struct {
	PaginationInfo PaginationInfo `json:"PaginationInfo"`
	PurlList       []Purl         `json:"PurlList"`
}

// AnalysisDependenciesParams defines parameters for AnalysisDependencies.
type AnalysisDependenciesParams struct {
	// PaginationSpec The pagination configuration for the query
	PaginationSpec *PaginationSpec `form:"PaginationSpec,omitempty" json:"PaginationSpec,omitempty"`

	// Sort The sort order of the packages
	//   * 'frequency' - The packages with the highest number of dependents
	//   * 'scorecard' - The packages with the lowest OpenSSF scorecard score
	Sort AnalysisDependenciesParamsSort `form:"sort" json:"sort"`
}

// AnalysisDependenciesParamsSort defines parameters for AnalysisDependencies.
type AnalysisDependenciesParamsSort string

// RetrieveDependenciesParams defines parameters for RetrieveDependencies.
type RetrieveDependenciesParams struct {
	// PaginationSpec The pagination configuration for the query
	PaginationSpec *PaginationSpec `form:"PaginationSpec,omitempty" json:"PaginationSpec,omitempty"`

	// Purl the purl of the dependent package
	Purl string `form:"purl" json:"purl"`
}